requirements.txt
boto3==1.9.229
botocore==1.12.229
Django==2.2.4
django-cleanup==4.0.0
django-currentuser==0.4.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-storages==1.7.2
django-widget-tweaks==1.4.5
docutils==0.15.2
jmespath==0.9.4
Pillow==6.1.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
